+++ src/main/java/tech/sirwellington/alchemy/annotations/access/package-info.java
@@ -14,6 +14,6 @@
  * limitations under the License.
  */
 /**
- * Annotations that document the application and use of Design Patterns.
+ * Annotations for documenting expectations about access to data or functionality.
  */
-package sir.wellington.alchemy.annotations.patterns;
+package tech.sirwellington.alchemy.annotations.access;
