+++ src/main/java/tech/sirwellington/alchemy/annotations/arguments/NonEmpty.java
@@ -13,29 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sir.wellington.alchemy.annotations.concurrency;
+
+
+package tech.sirwellington.alchemy.annotations.arguments;
 
 import java.lang.annotation.Documented;
-import static java.lang.annotation.ElementType.CONSTRUCTOR;
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.LOCAL_VARIABLE;
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.ElementType.PARAMETER;
 import java.lang.annotation.Target;
+import java.util.Collection;
 
 /**
- * Designate the use of an Object or variable as Thread-Safe. This means that no special precautions
- * must be taken to handle concurrency issues while using this object. Usually Objects that are
- * immutable or handle synchronization internally can be marked as {@link ThreadSafe}.
+ * Specifies that the argument or field CANNOT be {@code null} or empty. This applies to
+ * {@linkplain String Strings}, {@code Arrays}, and {@link Collection Collections}.
  *
- * @see ThreadUnsafe
- * @see Immutable
- * 
  * @author SirWellington
  */
 @Documented
-@Target({ TYPE, FIELD, CONSTRUCTOR, LOCAL_VARIABLE, METHOD})
-public @interface ThreadSafe
+@Target({PARAMETER, FIELD, LOCAL_VARIABLE})
+public @interface NonEmpty 
 {
 
 }
