+++ src/main/java/tech/sirwellington/alchemy/annotations/arguments/NonNull.java
@@ -15,24 +15,25 @@
  */
 
 
-package sir.wellington.alchemy.annotations.arguments;
+package tech.sirwellington.alchemy.annotations.arguments;
 
 import java.lang.annotation.Documented;
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.LOCAL_VARIABLE;
+import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.PARAMETER;
 import java.lang.annotation.Target;
-import java.util.Collection;
 
 /**
- * Specifies that the argument or field CANNOT be {@code null} or empty. This applies to
- * {@linkplain String Strings}, {@code Arrays}, and {@link Collection Collections}.
+ * Specifies that the argument or field CANNOT be {@code null}.
  *
+ * @see Nullable
+ * 
  * @author SirWellington
  */
 @Documented
-@Target({PARAMETER, FIELD, LOCAL_VARIABLE})
-public @interface NonEmpty 
+@Target({PARAMETER, FIELD, LOCAL_VARIABLE, METHOD})
+public @interface NonNull 
 {
 
 }
