+++ src/main/java/tech/sirwellington/alchemy/annotations/arguments/package-info.java
@@ -13,25 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-
-package sir.wellington.alchemy.annotations.arguments;
-
-import java.lang.annotation.Documented;
-import static java.lang.annotation.ElementType.FIELD;
-import static java.lang.annotation.ElementType.LOCAL_VARIABLE;
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.PARAMETER;
-import java.lang.annotation.Target;
-
 /**
- * Specifies that {@code null} is an acceptable value for the argument or field.
- *
- * @author SirWellington
+ * Annotations for documenting expectations about arguments and fields.
  */
-@Documented
-@Target({PARAMETER, FIELD, LOCAL_VARIABLE, METHOD})
-public @interface Nullable 
-{
-
-}
+package tech.sirwellington.alchemy.annotations.arguments;
