+++ src/main/java/tech/sirwellington/alchemy/annotations/concurrency/Immutable.java
@@ -13,9 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-
-package sir.wellington.alchemy.annotations.concurrency;
+package tech.sirwellington.alchemy.annotations.concurrency;
 
 import java.lang.annotation.Documented;
 import static java.lang.annotation.ElementType.FIELD;
@@ -23,19 +21,17 @@ import static java.lang.annotation.ElementType.LOCAL_VARIABLE;
 import static java.lang.annotation.ElementType.TYPE;
 import java.lang.annotation.Target;
 
-
 /**
- * Used to indicate that a variable or class is not designed to be immutable. That is, its
- * underlying can be changed once set.
+ * Used to indicate that a variable or class is designed to be immutable. Once it is created and
+ * set, it's underlying state cannot be changed.
  *
- * @see Immutable
+ * @see Mutable
  * 
  * @author SirWellington
  */
-
 @Documented
 @Target({ TYPE, FIELD, LOCAL_VARIABLE})
-public @interface Mutable
+public @interface Immutable
 {
 
 }
