+++ src/main/java/tech/sirwellington/alchemy/annotations/concurrency/package-info.java
@@ -13,23 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sir.wellington.alchemy.annotations.patterns;
-
-import java.lang.annotation.Documented;
-import static java.lang.annotation.ElementType.TYPE;
-import java.lang.annotation.Target;
-
 /**
- * Apply to Classes that apply the Singleton Design Pattern.
- *
- * @see
- * <a href="http://www.oodesign.com/singleton-pattern.html">http://www.oodesign.com/singleton-pattern.html</a>
- *
- * @author SirWellington
+ * Annotations documenting concepts and issues relating to Concurrency.
  */
-@Documented
-@Target(TYPE)
-public @interface SingletonPattern
-{
-
-}
+package tech.sirwellington.alchemy.annotations.concurrency;
