+++ src/main/java/tech/sirwellington/alchemy/annotations/patterns/FluidAPIPattern.java
@@ -13,7 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package tech.sirwellington.alchemy.annotations.patterns;
+
+import java.lang.annotation.Documented;
+import static java.lang.annotation.ElementType.TYPE;
+import java.lang.annotation.Target;
+
 /**
- * Annotations documenting concepts and issues relating to Concurrency.
+ * Used to document the application or use of the Fluid API Style, explained
+ * <a href="https://en.wikipedia.org/wiki/Fluent_interface">here</a>.
+ *
+ * @see
+ * <a href="https://en.wikipedia.org/wiki/Fluent_interface">https://en.wikipedia.org/wiki/Fluent_interface</a>
+ * @author SirWellington
  */
-package sir.wellington.alchemy.annotations.concurrency;
+@Documented
+@Target(TYPE)
+public @interface FluidAPIPattern
+{
+
+}
